Separate methods in different files for reduce lines number (as requested by linters)
