+++ my_count.rb
@@ -1,0 +1,14 @@
+module Enumerable
+  # print out the number of elements
+  def my_count(ele = nil)
+    counter = 0
+    if block_given? # to prevent block from throwing errors
+        my_each { |n| counter += 1 if yield(n) }
+    elsif ele
+        my_each { |n| counter += 1 if n == ele }
+    else
+      counter = self.length
+    end
+    counter
+  end
+end
