+++ my_inject_and_multiply_els.rb
@@ -1,0 +1,14 @@
+module Enumerables
+  def my_inject(*input) # to sum-up
+    return enum_for unless block_given? #to prevent block from throwing errors
+    argument = input + self
+    return nil if argument.length == 0
+    return argument[0] if argument.length == 1
+    output = argument[0]
+    argument[1..argument.length - 1].my_each {|num| output = yield(output, num)}
+    return output
+  end
+end
+def multiply_els(input) # for multiplication
+  input.my_inject { |k, n| k * n}
+end
